# api-docs/openapi-device-base-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "设备基础数据业务模块 API 文档",
    "description": "weili-iot-portal 业务模块中设备基础数据（device-base）模块的 REST API 接口文档。\n\n功能覆盖设备基础信息、设备配置、设备型号、设备类型、组织单元以及设备关系等主数据维护。\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/api/device-base/v1",
      "description": "设备基础数据 API 基础路径"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/device-types/{id}": {
      "get": {
        "tags": [
          "设备类型"
        ],
        "summary": "查询设备类型详情",
        "operationId": "getDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "设备类型"
        ],
        "summary": "更新设备类型",
        "operationId": "updateDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTypeUpdateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "设备类型"
        ],
        "summary": "删除设备类型",
        "operationId": "deleteDeviceType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "设备类型ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/device-types/page": {
      "post": {
        "tags": [
          "设备类型"
        ],
        "summary": "分页查询设备类型",
        "operationId": "pageDeviceType",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTypeQueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultPageDeviceTypeVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/organization-units/{id}": {
      "get": {
        "tags": [
          "组织单元"
        ],
        "summary": "查询组织单元详情",
        "operationId": "getOrganizationUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "组织单元ID"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultOrganizationUnitVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "组织单元"
        ],
        "summary": "更新组织单元",
        "operationId": "updateOrganizationUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "组织单元ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUnitUpdateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultOrganizationUnitVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "组织单元"
        ],
        "summary": "删除组织单元",
        "operationId": "deleteOrganizationUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "组织单元ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/organization-units/page": {
      "post": {
        "tags": [
          "组织单元"
        ],
        "summary": "分页查询组织单元",
        "operationId": "pageOrganizationUnit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUnitQueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultPageOrganizationUnitVO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommonResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码（200 表示成功）"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "时间戳"
          }
        }
      },
      "CommonResultBoolean": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "PageResult": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pageNo": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "DeviceTypeVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "设备类型ID（主键）"
          },
          "typeCode": {
            "type": "string",
            "description": "类型编码（唯一标识）"
          },
          "typeName": {
            "type": "string",
            "description": "类型名称（显示名称）"
          },
          "parentTypeId": {
            "type": "string",
            "description": "父类型ID（用于树形结构，根节点为null）"
          },
          "parentTypeName": {
            "type": "string",
            "description": "父类型名称（冗余字段，便于展示）"
          },
          "level": {
            "type": "integer",
            "description": "层级（1-主类型，2-子类型）"
          },
          "category": {
            "type": "string",
            "description": "分类（如：CNC、PLC等）"
          },
          "description": {
            "type": "string",
            "description": "类型描述（详细说明）"
          },
          "icon": {
            "type": "string",
            "description": "图标（图标URL或图标代码）"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "自定义字段（JSON对象，存储扩展属性）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（true-启用，false-停用）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（数字越小越靠前）"
          },
          "createdTime": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（毫秒时间戳）"
          },
          "updatedTime": {
            "type": "integer",
            "format": "int64",
            "description": "更新时间（毫秒时间戳）"
          }
        }
      },
...
      "DeviceTypeUpdateReq": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "设备类型ID（必填，用于更新）"
          },
          "typeName": {
            "type": "string",
            "description": "类型名称（可选）"
          },
          "description": {
            "type": "string",
            "description": "类型描述（可选）"
          },
          "icon": {
            "type": "string",
            "description": "图标（可选）"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "自定义字段（可选，JSON对象）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（可选）"
          }
        }
      },
      "DeviceTypeQueryReq": {
        "type": "object",
        "required": [
          "pageNo",
          "pageSize"
        ],
        "properties": {
          "typeCodeLike": {
            "type": "string",
            "description": "类型编码模糊查询（可选）"
          },
          "typeNameLike": {
            "type": "string",
            "description": "类型名称模糊查询（可选）"
          },
          "parentTypeId": {
            "type": "string",
            "description": "父类型ID（可选，查询指定父类型下的子类型）"
          },
          "level": {
            "type": "integer",
            "description": "层级查询（可选，1-主类型，2-子类型）"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "分类列表（可选，支持多选）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选，true-仅查询启用，false-仅查询停用，不传-不过滤）"
          },
          "pageNo": {
            "type": "integer",
            "minimum": 1,
            "description": "页码（必填，从1开始）"
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "description": "每页条数（必填，1-200）"
          },
          "sortBy": {
            "type": "string",
            "description": "排序字段（可选）"
          },
          "sortDirection": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ],
            "description": "排序方向（可选，ASC-升序，DESC-降序）"
          }
        }
      },
      "CommonResultDeviceTypeVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/DeviceTypeVO"
              }
            }
          }
        ]
      },
      "CommonResultPageDeviceTypeVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PageResult"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "records": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/DeviceTypeVO"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
      "OrganizationUnitVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "组织单元ID（主键）"
          },
          "unitCode": {
            "type": "string",
            "description": "组织单元编码（唯一标识）"
          },
          "unitName": {
            "type": "string",
            "description": "组织单元名称（显示名称）"
          },
          "unitType": {
            "type": "string",
            "description": "组织单元类型（FACTORY-厂区、WORKSHOP-车间、LINE-产线）"
          },
          "parentId": {
            "type": "string",
            "description": "父组织单元ID（用于树形结构，根节点为null）"
          },
          "parentName": {
            "type": "string",
            "description": "父组织单元名称（冗余字段，便于展示）"
          },
          "level": {
            "type": "integer",
            "description": "层级（1-厂区，2-车间，3-产线）"
          },
          "path": {
            "type": "string",
            "description": "路径（组织单元的全路径，如：/厂区1/车间1/产线1）"
          },
          "description": {
            "type": "string",
            "description": "描述（组织单元详细说明）"
          },
          "location": {
            "type": "string",
            "description": "位置（组织单元的地理位置或位置描述）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（true-启用，false-停用）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（数字越小越靠前）"
          },
          "createdTime": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（毫秒时间戳）"
          },
          "updatedTime": {
            "type": "integer",
            "format": "int64",
            "description": "更新时间（毫秒时间戳）"
          },
          "createdBy": {
            "type": "string",
            "description": "创建人（用户ID或用户名）"
          },
          "updatedBy": {
            "type": "string",
            "description": "更新人（用户ID或用户名）"
          }
        }
      },
...
      "OrganizationUnitUpdateReq": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "组织单元ID（必填，用于更新）"
          },
          "unitName": {
            "type": "string",
            "description": "组织单元名称（可选）"
          },
          "description": {
            "type": "string",
            "description": "描述（可选）"
          },
          "location": {
            "type": "string",
            "description": "位置（可选）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选）"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序（可选）"
          }
        }
      },
      "OrganizationUnitQueryReq": {
        "type": "object",
        "required": [
          "pageNo",
          "pageSize"
        ],
        "properties": {
          "unitCodeLike": {
            "type": "string",
            "description": "组织单元编码模糊查询（可选）"
          },
          "unitNameLike": {
            "type": "string",
            "description": "组织单元名称模糊查询（可选）"
          },
          "unitTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "组织单元类型列表（可选，支持多选，如：[\"FACTORY\", \"WORKSHOP\"]）"
          },
          "parentId": {
            "type": "string",
            "description": "父组织单元ID（可选，查询指定父单元下的子单元）"
          },
          "level": {
            "type": "integer",
            "description": "层级查询（可选，1-厂区，2-车间，3-产线）"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用（可选，true-仅查询启用，false-仅查询停用，不传-不过滤）"
          },
          "pageNo": {
            "type": "integer",
            "minimum": 1,
            "description": "页码（必填，从1开始）"
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "description": "每页条数（必填，1-200）"
          },
          "sortBy": {
            "type": "string",
            "description": "排序字段（可选）"
          },
          "sortDirection": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ],
            "description": "排序方向（可选，ASC-升序，DESC-降序）"
          }
        }
      },
      "CommonResultOrganizationUnitVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/OrganizationUnitVO"
              }
            }
          }
        ]
      },
      "CommonResultPageOrganizationUnitVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonResult"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PageResult"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "records": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/OrganizationUnitVO"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未授权访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CommonResult"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证"
      }
    }
  }
}